1992 - clean up useOnChangeNodeValue
