+++ src/client/pages/Section/DataTable/Table/RowData/Cell/hooks/useOnChange/useAffectedNodes.ts
@@ -1,0 +1,91 @@
+import { useCallback } from 'react'
+
+import { Objects } from 'utils/objects'
+
+import { CountryIso } from 'meta/area'
+import { VariableCache } from 'meta/assessment/metaCache'
+import { AssessmentMetaCaches } from 'meta/assessment/metaCaches'
+import { NodeValue } from 'meta/assessment/node'
+import { RowCache } from 'meta/assessment/rowCache'
+import { RecordAssessmentDatas } from 'meta/data'
+import { ExpressionEvaluator } from 'meta/expressionEvaluator'
+
+import { useAssessment, useCycle } from 'client/store/assessment'
+import { useCountryRouteParams } from 'client/hooks/useRouteParams'
+
+import { Props } from './types'
+
+const useAffectedNodes = (props: Props) => {
+  const { col, data, row, table } = props
+
+  const { countryIso } = useCountryRouteParams<CountryIso>()
+  const assessment = useAssessment()
+  const cycle = useCycle()
+
+  return useCallback(
+    (value: NodeValue): VariableCache[] => {
+      const tableName = table.props.name
+      const { variableName } = row.props
+      const { colName } = col.props
+
+      // Future data contains the updated value and is used to evaluate the enablers
+      const futureData = RecordAssessmentDatas.updateDatum({
+        assessmentName: assessment.props.name,
+        cycleName: cycle.name,
+        countryIso,
+        tableName,
+        colName,
+        variableName,
+        data: Objects.cloneDeep(data),
+        value,
+      })
+
+      const enablerDependants = AssessmentMetaCaches.getEnablersDependants({
+        assessment,
+        cycle,
+        tableName,
+        variableName,
+      })
+
+      return enablerDependants.filter((dep) => {
+        const depRow = table.rows?.find((r) => r.props.variableName === dep.variableName)
+        const depCol = depRow?.cols?.find((c) => c.props.colName === dep.colName)
+        if (!depRow || !depCol) return false
+        const enableIf = depCol.props.enableIf?.[cycle.uuid]
+        if (!enableIf) return false
+
+        const rowCache: RowCache = {
+          ...depRow,
+          tableName,
+          sectionName: props.sectionName,
+        }
+
+        const enabled = ExpressionEvaluator.evalFormula<boolean>({
+          assessment,
+          assessments: { [assessment.props.name]: assessment },
+          countryIso,
+          cycle,
+          data: futureData,
+          colName: depCol.props.colName,
+          row: rowCache,
+          formula: enableIf,
+        })
+
+        const depValue = RecordAssessmentDatas.getNodeValue({
+          assessmentName: assessment.props.name,
+          cycleName: cycle.name,
+          data,
+          countryIso,
+          tableName,
+          colName: depCol.props.colName,
+          variableName: depRow.props.variableName,
+        })
+
+        return !enabled && !!depValue && !!depValue.raw
+      })
+    },
+    [assessment, col, countryIso, cycle, data, props.sectionName, row, table]
+  )
+}
+
+export default useAffectedNodes
